+++ ChessAndroid/app/src/main/cpp/chess/data/pieces/BishopPiece.h
@@ -13,5 +13,5 @@ public:
 
 private:
 	void calculateMoves(Pos &pos, std::vector<Pos> &moves, const Board &board) const override;
-	float evaluatePiece(const int x, const int y) const override;
+    int evaluatePiece(const int x, const int y) const override;
 };
