+++ ChessAndroid/app/src/main/cpp/chess/data/pieces/Piece.h
@@ -30,7 +30,7 @@ public:
 	virtual ~Piece() = default;
 
 	const std::vector<Pos> getPossibleMoves(Pos pos, const Board &board) const;
-	float getPoints(const int x, const int y) const;
+	int getPoints(const int x, const int y) const;
 	bool hasSameColor(const Piece &other) const;
 
 	bool operator==(const Piece &other) const;
@@ -39,5 +39,5 @@ protected:
 	bool isMaximising() const;
 
 	virtual void calculateMoves(Pos &pos, std::vector<Pos> &moves, const Board &board) const = 0;
-	virtual float evaluatePiece(const int x, const int y) const = 0;
+	virtual int evaluatePiece(const int x, const int y) const = 0;
 };
